Rename result file
